Big refactor and clean up
improved error handler

refactor of server.ts to come.
