+++ backend/ormconfig.json
@@ -1,18 +1,16 @@
 {
-   "type": "postgres",
-   "host": "localhost",
-   "port": 5432,
-   "username": "postgres",
-   "password": "rootpassword",
-   "database": "ezybucks",
-   "synchronize": true,
-   "logging": false,
-   "entities": [
-      "dist/entity/**/*.js"
-   ],
-   "cli": {
-      "entitiesDir": "backend/src/entity",
-      "migrationsDir": "backend/src/migration",
-      "subscribersDir": "backend/src/subscriber"
-   }
+    "type": "postgres",
+    "host": "localhost",
+    "port": 5432,
+    "username": "postgres",
+    "password": "rootpassword",
+    "database": "ezybucks",
+    "synchronize": true,
+    "logging": false,
+    "entities": ["dist/entities/**/*.js"],
+    "cli": {
+        "entitiesDir": "backend/src/entity",
+        "migrationsDir": "backend/src/migration",
+        "subscribersDir": "backend/src/subscriber"
+    }
 }
