+++ backend/src/error/HttpException.ts
@@ -1,0 +1,12 @@
+class HttpException extends Error {
+    public status: number;
+    public message: string;
+
+    constructor(status: number, message: string) {
+        super(message);
+        this.status = status;
+        this.message = message;
+    }
+}
+
+export default HttpException;
