+++ backend/src/error/error.middleware.ts
@@ -1,0 +1,19 @@
+import { NextFunction, Request, Response } from 'express';
+import HttpException from './HttpException';
+
+function errorMiddleware(
+    error: HttpException,
+    req: Request,
+    res: Response,
+    next: NextFunction
+) {
+    const status = error.status || 500;
+    const message = error.message || 'Something went wrong! :(';
+
+    res.status(status).json({
+        status,
+        message
+    });
+}
+
+export default errorMiddleware;
