--- backend/src/error/responseError.interface.ts
@@ -1,3 +1,0 @@
-interface IResponseError extends Error {
-    status?: number;
-}
